+++ src/every.ts
@@ -2,19 +2,27 @@ import hasOwn from "./has-own";
 import { Extra, FilterCallback, Key } from "./types";
 import { wrapFilterCallback } from "./wrap-callback";
 
-function every<K extends Key, V, E extends Extra, TH = any>(
+function every<V, K extends Key, E extends Extra, TH = any>(
   this: TH,
   object: Record<K, V>,
-  callback: FilterCallback<K, V, E, TH>,
+  callback: FilterCallback<V, K, E, TH>,
   ...extra: E
 ): boolean;
-function every<K extends Key, V, E extends Extra, TH = any>(
+
+function every<V, K extends Key, TH = any>(
   this: TH,
   object: Record<K, V>,
-  callback: FilterCallback<K, V, E, TH>,
+  callback: FilterCallback<V, K, Extra, TH>,
+  ...extra: Extra
+): boolean;
+
+function every<V, K extends Key, E extends Extra, TH = any>(
+  this: TH,
+  object: Record<K, V>,
+  callback: FilterCallback<V, K, E, TH>,
 ): boolean {
 
-  const cb = wrapFilterCallback<K, V, E, TH, any>(
+  const cb = wrapFilterCallback<V, K, E, TH, any>(
     callback,
     arguments,
   );