+++ src/includes.ts
@@ -1,10 +1,10 @@
 import keyOf from "./key-of";
 import { Key } from "./types";
 
-function includes(
-  object: Record<Key, any>,
-  value: any,
-): boolean {
+function includes(object: {}, value: any): false;
+function includes(object: Record<Key, any>, value: any): boolean;
+
+function includes(object: Record<Key, any>, value: any): boolean {
   return keyOf(object, value) !== null;
 }