+++ src/key-of.ts
@@ -1,10 +1,10 @@
 import hasOwn from "./has-own";
 import { Key } from "./types";
 
-function keyOf<K extends Key>(
-  object: Record<K, any>,
-  value: any,
-): K | null {
+function keyOf(object: {}, value: any): null;
+function keyOf<K extends Key>(object: Record<K, any>, value: any): K | null;
+
+function keyOf<K extends Key>(object: Record<K, any>, value: any): K | null {
 
   for (const key in object) {
     if (