+++ src/last-key-of.ts
@@ -1,10 +1,10 @@
 import hasOwn from "./has-own";
 import { Key } from "./types";
 
-function lastKeyOf<K extends Key>(
-  object: Record<K, any>,
-  value: any,
-): K | null {
+function lastKeyOf(object: {}, value: any): null;
+function lastKeyOf<K extends Key>(object: Record<K, any>, value: any): K | null;
+
+function lastKeyOf<K extends Key>(object: Record<K, any>, value: any): K | null {
 
   let result: K | null = null;