+++ src/types.ts
@@ -2,8 +2,8 @@ export type Key = Extract<keyof any, string | number>;
 export type Extra = any[];
 
 export type MapCallback<
-  K extends Key,
   V,
+  K extends Key,
   E extends Extra,
   TH = any,
   R = any> = (
@@ -14,29 +14,29 @@ export type MapCallback<
   ) => R;
 
 export type ForEachCallback<
-  K extends Key,
   V,
+  K extends Key,
   E extends Extra,
-  TH = any> = MapCallback<K, V, E, TH, void>;
+  TH = any> = MapCallback<V, K, E, TH, void>;
 
 export type FilterCallback<
-  K extends Key,
   V,
+  K extends Key,
   E extends Extra,
-  TH = any> = MapCallback<K, V, E, TH, any>;
+  TH = any> = MapCallback<V, K, E, TH, any>;
 
 export type WrappedFilterCallback<
-  K extends Key,
   V,
+  K extends Key,
   TH = any,
   R = any> = (thisArg: TH, object: Record<K, V>, key: K) => R;
 
 export type ReduceCallback<
-  K extends Key,
   V,
+  K extends Key,
   E extends Extra,
-  R = any,
-  TH = any> = (
+  TH = any,
+  R = any> = (
     this: TH,
     prev: R,
     value: V,
@@ -45,7 +45,7 @@ export type ReduceCallback<
   ) => R;
 
 export type WrappedReduceCallback<
-  K extends Key,
   V,
+  K extends Key,
   TH = any,
   R = any> = (thisArg: TH, object: Record<K, V>, key: K, result: R | undefined) => R;