+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * improved documentation
 * performance improvements for none & 1 extra arguments
 * dependencies updated
+* types refactoring
 
 # 1.1.13 (01-06-2019)